feat: new methods for Depoimentos and DataFile

+++ src/main/java/br/com/pedroargentati/viagens_api/repository/DepoimentosRepository.java
@@ -3,6 +3,6 @@ package br.com.pedroargentati.viagens_api.repository;
 import br.com.pedroargentati.viagens_api.model.Depoimentos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepoimentosRepository extends JpaRepository<Depoimentos, Long> {
+public interface DepoimentosRepository extends JpaRepository<Depoimentos, Integer> {
 
 }
